requirements.txt
fastapi
uvicorn
jinja2
python-multipart
starlette
deepface
tf-keras
opencv-python
numpy
python-dotenv
<<<<<<< Updated upstream
itsdangerous
=======
httpx
>>>>>>> Stashed changes
